[FIX] sale_subscription/saas~12.4.1.1: correct column type
Otherwise the ORM takes lot of time correcting it
```
2021-11-02 19:34:10,598 2229 INFO db_42897 odoo.modules.registry: module sale_subscription: creating or updating database tables
2021-11-02 19:50:26,714 2229 INFO db_42897 odoo.modules.loading: loading sale_subscription/security/sale_subscription_security.xml
```

upg-42897
opw-2672031

Part-of: odoo/upgrade#3039

+++ migrations/sale_subscription/saas~12.4.1.1/pre-10-account-pocalypse.py
@@ -6,4 +6,4 @@ def migrate(cr, version):
     util.create_column(cr, 'account_move_line', 'subscription_id', 'int4')
     util.create_column(cr, 'account_move_line', 'subscription_start_date', 'date')
     util.create_column(cr, 'account_move_line', 'subscription_end_date', 'date')
-    util.create_column(cr, 'account_move_line', 'subscription_mrr', 'float')
+    util.create_column(cr, 'account_move_line', 'subscription_mrr', 'numeric')
